requirements.txt
asgiref==3.7.2
Django==4.2.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
Pillow==9.5.0
sqlparse==0.4.4
typing_extensions==4.6.2
